+++ src/main/java/online/kingdomkeys/kingdomkeys/advancements/ModAdvancements.java
@@ -21,6 +21,7 @@ public class ModAdvancements
         public static final DeferredHolder<CriterionTrigger<?>, KKChoiceTrigger> CHOICE_MADE = TRIGGERS.register("choice_made", KKChoiceTrigger::new);
         public static final DeferredHolder<CriterionTrigger<?>, KKKeybladeLevelTrigger> KEYBLADE_LEVEL = TRIGGERS.register("keyblade_level", KKKeybladeLevelTrigger::new);
         public static final DeferredHolder<CriterionTrigger<?>, KKCraftProfileHeadTrigger> CRAFT_PROFILE_HEAD = TRIGGERS.register("craft_profile_head", KKCraftProfileHeadTrigger::new);
+        public static final DeferredHolder<CriterionTrigger<?>, KKAllAdvancementsTrigger> ALL_ADVANCEMENTS = TRIGGERS.register("all_advancements", KKAllAdvancementsTrigger::new);
 
         public static void triggerMunnyReached(ServerPlayer player, int currentMunny) {
             MUNNY_REACHED.get().trigger(player, currentMunny);
@@ -45,39 +46,14 @@ public class ModAdvancements
         public static void triggerCraftProfileHead(ServerPlayer player, String profileName) {
             CRAFT_PROFILE_HEAD.get().trigger(player, profileName);
         }
-        /*public static final Id FULL_FERTILIZER = registerGeneric("full_fertilizer");
-        public static final Id LAVA_LAMP = registerGeneric("lava_lamp");
-        public static final Id ROTTEN_COMPOST_KILL = registerGeneric("rotten_compost_kill");
-        public static final Id PRESENT_DAY = registerGeneric("present_day");
-        public static final Id EAT_ROTTEN_FOOD = registerGeneric("eat_rotten_food");
-        public static final Id PERFECTLY_FORGED = registerGeneric("perfectly_forged");
-        public static final Id FULL_NUTRITION = registerGeneric("full_nutrition");
-        public static final Id MAX_WINDMILL = registerGeneric("max_windmill");
-        public static final Id BASIN_POUR = registerGeneric("basin_pour");
-        public static final Id TABLE_POUR = registerGeneric("table_pour");
 
-        public static final Id1<BlockState> CHISELED = registerBlock("chiseled");
-        public static final Id1<BlockState> LIT = registerBlock("lit");
-        public static final Id1<BlockState> ROCK_ANVIL = registerBlock("rock_anvil");
-        public static final Id1<BlockState> FIREPIT_CREATED = registerBlock("firepit_created");
-
-        public static final Id1<Entity> HOOKED_ENTITY = registerEntity("hooked_entity");
-        public static final Id1<Entity> FED_ANIMAL = registerEntity("fed_animal");
-        public static final Id1<Entity> STAB_ENTITY = registerEntity("stab_entity");*/
-
-        /*public static Id1<BlockState> registerBlock(String name)
-        {
-            return Id1.of(TRIGGERS.register(name, BlockActionTrigger::new), BlockActionTrigger::trigger);
-        }*/
+        public static void triggerAllAdvancements(ServerPlayer player) {
+            ALL_ADVANCEMENTS.get().trigger(player);
+        }
 
         public static Id registerGeneric(String name) {
             return new Id(TRIGGERS.register(name, KKLevelUpTrigger::new));
         }
-
-       /* public static Id1<Entity> registerEntity(String name)
-        {
-            return Id1.of(TRIGGERS.register(name, EntityActionTrigger::new), EntityActionTrigger::trigger);
-        }*/
 
         public record Id(DeferredHolder<CriterionTrigger<?>, KKLevelUpTrigger> holder) {
             public void trigger(ServerPlayer player, int level) {